fixes segfault when moving camera too quickly, reorganize files
