+++ scenes/triangle.rt
@@ -1,8 +1,8 @@
 A	0				0,255,0
 C	0,0,0			0,0,-1			90
 L	0,0,0			0.7				255,255,255
-tr	0,0,5			-10,10,6		10,10,7					255,0,0
-tr	0,0,-5			-10,10,-6		10,10,-7				255,0,0
+tr	0,0,5			-10,10,6		10,10,7					255,255,0
+tr	0,0,-5			-10,10,-6		10,10,-7				255,255,0
 pl	-100,0,0		1,0,0			255,0,0
 pl	100,0,0			1,0,0			255,0,0
 pl	0,0,-100		0,0,-1			0,0,255
